chore: Remove player from assignment lists
This commit modifies the code in the `src/routes/assignments/+page.svelte` file to add a new function `removePlayer` that removes a player from the assignment lists. The function is triggered when the "Remove" button is clicked for a player. This change improves the functionality of the assignments page by allowing users to remove players from the lists.