chore: Add watermark option to MessagePreview
The code changes in `src/lib/AAR/MessagePreview.svelte` add a new checkbox option for adding a watermark to the final message. When the checkbox is checked, the watermark "Made with [med-tools.space](https://med-tools.space)" is appended to the message. This change enhances the MessagePreview component by allowing users to easily add a watermark to their messages, providing attribution to the med-tools.space website.