Added sorting and modification abilities to the teams
This is including the functionality to
- delete teams
- move them to the first position in one click
- move them to the last position in one click
- change the team lead and the status of the team
- create teams in the "mustering" state
- keep states between visits
- change the number of the team without causing problems

The code for drag-n-drop was refactored to rerender the DOM when updating the order to preserve events within the team-card.

Teams will now pick the lowest number available to them (being limited from 1-999) instead of picking the highest number plus one.

Added alerts to indicate what is happening.

Added a bit of documentation, will have to finish that later, though.