Added the ability to generate the Dispatch Feed
Included the default format of the Dispatch Feed.
This includes coloring, the hidden queue when only one team is active, the current timestamp and the generalized length of the team names when one team leader has a longer name than the other ones.

Already included is the ability to add comments to a team, which will be displayed after the status.